apps/rag/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
psycopg2-binary==2.9.10
pgvector==0.3.6

# RAG + Embeddings
langchain==0.3.0
langchain-community==0.3.0
sentence-transformers==3.3.0
pymupdf==1.25.0
python-docx==1.1.2
groq==0.13.0

# Google APIs
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1

# Config
pydantic-settings==2.6.0
python-dotenv==1.0.1
